+++ demos/chinook/src/main/java/is/codion/demos/chinook/ui/ChinookAppPanel.java
@@ -54,8 +54,6 @@ import is.codion.swing.framework.ui.ReferentialIntegrityErrorHandling;
 import is.codion.swing.framework.ui.icon.FrameworkIcons;
 import is.codion.tools.swing.mcp.SwingMcpPlugin;
 
-import org.kordamp.ikonli.foundation.Foundation;
-
 import javax.swing.ButtonGroup;
 import javax.swing.JOptionPane;
 import javax.swing.JPanel;
@@ -184,7 +182,7 @@ public final class ChinookAppPanel extends EntityApplicationPanel<ChinookAppMode
 										.control(Control.builder()
 														.command(this::displayAnalytics)
 														.caption(bundle.getString("analytics"))
-														.icon(FrameworkIcons.instance().get(Foundation.GRAPH_PIE))
+														.icon(FrameworkIcons.instance().get("graph-pie"))
 														.build())
 										.separator()
 										.control(Control.builder()
@@ -245,7 +243,10 @@ public final class ChinookAppPanel extends EntityApplicationPanel<ChinookAppMode
 	public static void main(String[] args) throws CancelException {
 		String language = UserPreferences.get(LANGUAGE_PREFERENCES_KEY, Locale.getDefault().getLanguage());
 		Locale.setDefault(LANGUAGE_IS.equals(language) ? LOCALE_IS : LOCALE_EN);
-		FrameworkIcons.instance().add(Foundation.PLUS, Foundation.MINUS, Foundation.GRAPH_PIE);
+		FrameworkIcons icons = FrameworkIcons.instance();
+		icons.put("plus", ChinookAppPanel.class.getResource("plus.svg"));
+		icons.put("minus", ChinookAppPanel.class.getResource("minus.svg"));
+		icons.put("graph-pie", ChinookAppPanel.class.getResource("graph-pie.svg"));
 		Completion.COMPLETION_MODE.set(Completion.Mode.AUTOCOMPLETE);
 		EntityApplicationPanel.CACHE_ENTITY_PANELS.set(true);
 		EntityApplicationPanel.SQL_TRACING.set(true);